+++ src/parser/exprs/create.cpp
@@ -23,8 +23,8 @@ namespace flag {
 struct CreateTable : AstPush {
     flag::ExprFlags expr_flags;
     flag::ColumnFlags column_flags;
-    std::array<std::optional<size_t>, 2> column_values{};
-    std::optional<size_t> table_name_index;
+    std::array<std::optional<uint32_t>, 2> column_values{};
+    std::optional<uint32_t> table_name_index;
 
     CreateTable() {
         expr_flags = flag::ExprFlags::NextCreate;
@@ -37,7 +37,7 @@ struct CreateTable : AstPush {
         );
     }
 
-    inline std::optional<Errors> string_token(size_t i) {
+    inline std::optional<Errors> string_token(uint32_t i) {
         if (expr_flags == flag::ExprFlags::NextColumn) {
             if (column_flags == flag::ColumnFlags::NextName) {
                 column_flags = flag::ColumnFlags::NextType;
@@ -82,7 +82,7 @@ struct CreateTable : AstPush {
             NodeKind::ColumnName,
             NodeKind::ColumnType};
 
-        size_t parent_id = ast.size();
+        uint32_t parent_id = ast.size();
         add_child_id(
             0,
             push_ast<uint8_t>(
@@ -91,7 +91,7 @@ struct CreateTable : AstPush {
             )
         );
 
-        for (size_t i = 0; i < column_values.size();i++) {
+        for (uint32_t i = 0; i < column_values.size();i++) {
             if (!column_values[i])
                 break;
 
@@ -123,7 +123,7 @@ struct CreateTable : AstPush {
 std::expected<Ast, Errors> create_expr_node(const std::vector<Token>& tokens) {
     CreateTable status;
 
-    for (size_t i = 0; i < tokens.size(); i++) {
+    for (uint32_t i = 0; i < tokens.size(); i++) {
         switch (tokens[i].token_kind) {
             case TokenKind::CREATE: {
                 if (status.expr_flags == flag::ExprFlags::NextCreate) {
